notification/updater/requirements.txt
requests==2.22.0
SQLAlchemy==1.3.12
python-dotenv==0.10.3
SQLAlchemy-Utils==0.36.1
mysql-connector-python==8.0.19